Create ChromSCape project folder

Creates a project folder that will be recognizable by ChromSCape Shiny application.

create_project_folder(
  output_directory,
  analysis_name = "Analysis_1",
  ref_genome = c("hg38", "mm10")[1]
)

Arguments

output_directory
Path towards the directory to create the 'ChromSCape_Analyses' folder and the analysis subfolder. If this path already contains the 'ChromSCape_Analyses' folder, will only create the analysis subfolder.

analysis_name
Name of the analysis. Must only contain alphanumerical characters or '_'.

ref_genome
Reference genome, either 'hg38' or 'mm10'.

Value

Creates the project folder and returns the root of the project.

Examples

dir = tempdir()  
create_project_folder(output_directory = dir,
 analysis_name = "Analysis_1")

#> [1] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1"

list.dirs(file.path(dir))

#>  [1] "/tmp/RtmpjzpsXF"                                                            
#>  [2] "/tmp/RtmpjzpsXF/ChromSCape_analyses"                                        
#>  [3] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1"                             
#>  [4] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1/correlation_clustering"      
#>  [5] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1/correlation_clustering/Plots"
#>  [6] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1/coverage"                    
#>  [7] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1/Diff_Analysis_Gene_Sets"     
#>  [8] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1/Filtering_Normalize_Reduce"  
#>  [9] "/tmp/RtmpjzpsXF/ChromSCape_analyses/Analysis_1/peaks"                       
#> [10] "/tmp/RtmpjzpsXF/downlit"