Filter genes based on peak calling refined annotation

Filter genes based on peak calling refined annotation

filter_genes_with_refined_peak_annotation(
  refined_annotation,
  peak_distance,
  signific,
  over,
  under
)

Arguments

refined_annotation
A data.frame containing each gene distance to real peak

peak_distance
Minimum distance to an existing peak to accept a given gene

signific
Indexes of all significantly differential genes

over
Indexes of all significantly overexpressed genes

under
Indexes of all significantly underexpressed genes

Value

List of significantly differential, overexpressed and underexpressed genes close enough to existing peaks