Warning for raw_counts_to_sparse_matrix

Warning for raw_counts_to_sparse_matrix

warning_raw_counts_to_sparse_matrix(
  files_dir_list,
  file_type = c("scBAM", "scBED", "SparseMatrix"),
  peak_file = NULL,
  n_bins = NULL,
  bin_width = NULL,
  genebody = NULL,
  extendPromoter = 2500,
  verbose = TRUE,
  ref = "hg38"
)

Arguments

files_dir_list
A named character vector of directory containing the raw files

file_type
Input file(s) type(s) ('scBED','scBAM','SparseMatrix')

peak_file
A file containing genomic location of peaks (NULL)

n_bins
The number of bins to tile the genome (NULL)

bin_width
The size of bins to tile the genome (NULL)

genebody
Count on genes (body + promoter) ? (NULL)

extendPromoter
If counting on genes, number of base pairs to extend up or downstream of TSS (2500).

verbose
Verbose (TRUE)

ref
reference genome to use (hg38)

Value

Error or warnings if the input are not correct