Create a simulated single cell datamatrix & cell annotation

Create a simulated single cell datamatrix & cell annotation

create_scDataset_raw(
  cells = 300,
  features = 600,
  featureType = c("window", "peak", "gene"),
  sparse = TRUE,
  nsamp = 4,
  ref = "hg38",
  batch_id = factor(rep(1, nsamp))
)

Arguments

cells
Number of cells (300)

features
Number of features (600)

featureType
Type of feature (window)

sparse
Is matrix sparse ? (TRUE)

nsamp
Number of samples (4)

ref
Reference genome ('hg38')

batch_id
Batch origin (factor((1,1,1,1))

Value

A list composed of * mat : a sparse matrix following an approximation of the negative binomial law (adapted to scChIPseq) * annot : a data.frame of cell annotation * batches : an integer vector with the batch number for each cell

Examples

# Creating a basic sparse 600 genomic bins x 300 cells matrix and annotation
l = create_scDataset_raw()
head(l$mat)

#> 6 x 300 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 300 column names ‘sample_1_c1’, ‘sample_1_c2’, ‘sample_1_c3’ ... ]]

#>                                                                               
#> chr1_1_5147117         1   . . . . . 1 200 2  .   1   .  2   1 . 6 2   .   . 1
#> chr1_5147118_10294233  3   1 1 1 . . .   . . 56   1   .  .   1 . . .   1   2 .
#> chr1_10294234_15441350 1 301 . 2 . 1 2   . .  .   1   . 96   1 . . . 372   2 3
#> chr1_15441351_20588466 1   . 1 2 . . .   . .  1   3   1  1 373 . . 1   .   . .
#> chr1_20588467_25735583 2   . . 1 2 1 .   1 .  .   .   1  1 115 . . 1   . 300 .
#> chr1_25735584_30882699 .   . . . . . .   1 .  . 262 264  .   1 . 1 .   . 404 .
#>                                                                               
#> chr1_1_5147117           . 147   2   . 187   .   . . 314 . 4 223   .   .   1 1
#> chr1_5147118_10294233    .   .   .   .   .   .   1 .  46 . .   1 196   . 242 .
#> chr1_10294234_15441350   2   . 319   . 253   .   2 1   1 1 2   1   1   2   . .
#> chr1_15441351_20588466   2   .   1   1   2   1 196 1   . 1 .   .   . 110   . .
#> chr1_20588467_25735583 160   .   .   2   .   2   . . 105 . .   . 289   .   . 1
#> chr1_25735584_30882699 177   1   . 299   . 228   1 .   1 2 .   .   1   .   . .
#>                                                                             
#> chr1_1_5147117         2 . .   .   2 1 . .   .   1   . .   1   . . 143 . 156
#> chr1_5147118_10294233  . . . 335   . . . .   .   . 111 2   .   . .   1 .   1
#> chr1_10294234_15441350 . 1 .   2   . 2 . 6   .   1   . 2   1   . .   1 1   .
#> chr1_15441351_20588466 . 1 .   .   . . 1 .   . 205   1 .   1 140 .   2 .   .
#> chr1_20588467_25735583 . . .   .   2 . . .   .   1  36 1 222   1 1   . 1   .
#> chr1_25735584_30882699 . . .   1 300 . 1 . 108   2   . 1   .   . 1   2 2   .
#>                                                                               
#> chr1_1_5147117           1 . 1 .   1   2   . 2   . 2 1   .   1   .   1 . 6   .
#> chr1_5147118_10294233    . 3 . 1   .   . 161 .   1 1 .   .   1   3 270 . .   .
#> chr1_10294234_15441350   . . . . 100   1   . 1  62 . 1   .   3 260   . . .   .
#> chr1_15441351_20588466 113 . . .   . 349   . 2   . . .   . 396   6 112 1 .   .
#> chr1_20588467_25735583   . . . 1   .   .   1 . 212 . . 148   .   .   . 2 . 304
#> chr1_25735584_30882699   . . . 1   .   1   . .   . . .   .   1   .   . . .   .
#>                                                                               
#> chr1_1_5147117         .   . . 3   .   3 248   1 2 1   1 135   .   1   .   1 .
#> chr1_5147118_10294233  . 144 . .   .   .   .   2 . .   1   2   4   .   1   1 .
#> chr1_10294234_15441350 .   3 . . 203 287   .   . . . 132 113   1   .   .   . .
#> chr1_15441351_20588466 .   1 . .   .   .   . 130 . .   .   . 237   .  67   . 2
#> chr1_20588467_25735583 1   . 1 2   1   2   1  37 . .   .   2   . 137   2   . .
#> chr1_25735584_30882699 .   . . 1   .   .   .   2 1 .   1   1   .   1 258 120 1
#>                                                                             
#> chr1_1_5147117         1 2 . . 1 . . . .   1 . . . 1 . . 1 1   1   1   .   .
#> chr1_5147118_10294233  . . . . . . . 1 1   . 3 2 2 1 . 2 1 .   .   . 160 135
#> chr1_10294234_15441350 . . . 1 . . . . 1 246 . . . 1 . . . .   .   .   .   1
#> chr1_15441351_20588466 1 . 1 1 . 2 1 . .   . 1 2 . . . . 1 2 123   .   .   .
#> chr1_20588467_25735583 . . 2 1 . . . 1 1   . . 1 . . . . 2 .   . 370   .   2
#> chr1_25735584_30882699 2 . . . 1 . . . .   . . . 1 . 2 . 1 .   .   1   1   1
#>                                                                                
#> chr1_1_5147117         268   .   . 1 315   2 1  .   . 418 1   . 1   1 . . . 2 .
#> chr1_5147118_10294233    2 197   1 .   .   . . 51 103   1 .   . .   . . 1 . . 2
#> chr1_10294234_15441350   .   . 256 2   1   2 .  2   .   . .   . 1   1 . . . 2 .
#> chr1_15441351_20588466   .   .   . .   .   . 1  1   .   . . 144 1 840 . . . . .
#> chr1_20588467_25735583   1   1   2 .   1   . 1  .   .   1 .   2 1 291 . . 1 . .
#> chr1_25735584_30882699  60 453   . .   . 378 .  .   .   1 3   . .   . . . . . .
#>                                                                               
#> chr1_1_5147117         .  .   . . . 1 . 2 . 1 .  . 1 1   .   . . .   1   .   .
#> chr1_5147118_10294233  .  1   . 1 1 . . . . . .  . . .   1   . . 3   .   .   .
#> chr1_10294234_15441350 .  .   . 1 1 1 1 . . . . 32 . .   .   1 . 3 267   .   2
#> chr1_15441351_20588466 .  .   . . . . 2 . . . .  2 . .   .   2 . .   .   .   1
#> chr1_20588467_25735583 2  .   3 . 2 . 1 . 1 1 .  1 . 1 266 168 . .   2   . 111
#> chr1_25735584_30882699 1 93 208 1 2 . . . 1 . .  . . 3   .   . 1 .   . 177   3
#>                                                                                
#> chr1_1_5147117         . . .   .  32   . 1 1 . . . . 173   . 1 . 106 .  .   . .
#> chr1_5147118_10294233  . . .   .   .   . . . . . 1 .   1 234 . .   1 .  . 267 .
#> chr1_10294234_15441350 1 1 1 228   1   . . . . . . 1   . 106 . .   . .  .   . .
#> chr1_15441351_20588466 . . .   .   1  96 6 1 . . . .   1   . . .   . .  2   1 3
#> chr1_20588467_25735583 . 1 .   . 173   1 2 . . 2 1 .   .   1 1 1   . . 19   . 2
#> chr1_25735584_30882699 1 . .   .   . 442 1 1 . 1 2 .   .   1 3 .   . .  .   . 1
#>                                                                             
#> chr1_1_5147117           . 1 . . .  . 1 . .  .   1 254 . . 122 .   . . 117 .
#> chr1_5147118_10294233    . . . . 2  1 . 2 1 87 261   7 . . 165 1 233 .   1 .
#> chr1_10294234_15441350   . . . . .  . . . 2  .   .   . . 1   . .   . .   . .
#> chr1_15441351_20588466   . 1 2 . 2 69 1 2 .  .   1   2 . 2   . .   . .   1 .
#> chr1_20588467_25735583 172 1 . . .  . . 2 1  .   .   . . .   2 .   2 2   . 1
#> chr1_25735584_30882699   . . . 1 .  . . . .  2   .   . . .   . .   2 .   . 2
#>                                                                             
#> chr1_1_5147117           2 . 231 339 . 346 . 1 1 .   1 2 1 . 1 . .   . . . .
#> chr1_5147118_10294233  287 .   . 164 1   1 . . 1 1   . 1 . . . . .   1 . 1 .
#> chr1_10294234_15441350   1 .   .   . .   . . 1 . . 215 2 2 . . . 1   2 1 2 .
#> chr1_15441351_20588466   2 .   .   3 .   2 1 1 1 .   . . 2 . . 2 .   1 . . .
#> chr1_20588467_25735583   1 1   1   2 1   . . 1 1 .  42 . 1 1 . 1 1   . . . .
#> chr1_25735584_30882699   4 .   .   2 .   3 3 . . .   . . . . . . . 182 . . .
#>                                                                             
#> chr1_1_5147117           2 .  35 . 210 .   . . . . .   . . 1 .   2   .   1 .
#> chr1_5147118_10294233   21 . 137 .   . .   . . . . .   1 1 . .   . 225 151 .
#> chr1_10294234_15441350   1 1   3 1   . .   . . . . .   . . 1 .   .   .   . .
#> chr1_15441351_20588466 273 .   . 2   . .   1 . . . 2 152 . . .   1   1   . .
#> chr1_20588467_25735583   . .   1 1 115 . 272 . . . .   . . . 3   .   .   . 1
#> chr1_25735584_30882699   . .   1 1   . 2  32 1 2 . 2   1 . 2 . 288   .   . .
#>                                                                                
#> chr1_1_5147117           .  . 1 . .  70 .   . .   .  . . . 1 2 249   2 .  . . 1
#> chr1_5147118_10294233    .  . . . .   . . 390 1   .  . . . 3 .   .   2 . 42 2 .
#> chr1_10294234_15441350   .  . . . .   1 .   1 .   . 90 . . . 1   1   . .  . 1 .
#> chr1_15441351_20588466   . 15 . 1 .   . .   . . 371  1 . . . 1   . 258 3  . . .
#> chr1_20588467_25735583   .  3 . . 1   . .   . .   .  3 2 . 1 .   .   1 .  . . 3
#> chr1_25735584_30882699 140  . . . . 128 1   . .   .  . 2 . . 1   .   . .  . . .
#>                                                                               
#> chr1_1_5147117         1   1   1 1  . . 1 1 .   .   . . . . . 343 . . 12 .   .
#> chr1_5147118_10294233  1 296   . .  1 . . 1 .   1   . 2 . 1 .   . 1 .  1 1   .
#> chr1_10294234_15441350 .   .   . . 98 . . . . 423 204 . . . .   . . 2  . . 188
#> chr1_15441351_20588466 .   . 331 1  . . . 1 1   9   . . . . .   . . 1  . .   .
#> chr1_20588467_25735583 2   1   1 2  1 . . 1 1   1   . . . . 2   1 . 1  . 1 143
#> chr1_25735584_30882699 .   .   1 .  . . 2 . .   .   . 1 . . . 327 . 1  2 2   .
#>                                                                             
#> chr1_1_5147117         .   . . . . 3 .  1   . 2 . . .   . .   . 68 168 . . .
#> chr1_5147118_10294233  .   1 2 . 2 1 .  .   1 . 1 1 .   1 1   .  .   . . . 1
#> chr1_10294234_15441350 .   . . 4 . 1 .  .   . . . 1 .   . .   .  1   . . . .
#> chr1_15441351_20588466 .   3 . . . 2 .  . 476 . . . 2   . .   .  .   . . . .
#> chr1_20588467_25735583 .   . . 1 . . . 54   . . . . .   . 1   2  .   . 1 . .
#> chr1_25735584_30882699 . 209 . 4 . . .  .   2 1 2 . . 327 . 379  .   . . . .
#>                                       
#> chr1_1_5147117           1 .   .   1 2
#> chr1_5147118_10294233    1 .   .   . 1
#> chr1_10294234_15441350   1 . 191 325 2
#> chr1_15441351_20588466 320 .   .   . 1
#> chr1_20588467_25735583   2 .   .   . .
#> chr1_25735584_30882699   . 2   .   1 1

head(l$annot)

#>             barcode     cell_id sample_id batch_id total_counts
#> sample_1_c1     BC1 sample_1_c1  sample_1        1        11264
#> sample_1_c2     BC2 sample_1_c2  sample_1        1        14545
#> sample_1_c3     BC3 sample_1_c3  sample_1        1        13741
#> sample_1_c4     BC4 sample_1_c4  sample_1        1        10384
#> sample_1_c5     BC5 sample_1_c5  sample_1        1        12683
#> sample_1_c6     BC6 sample_1_c6  sample_1        1        11757

head(l$batches)

#> [1] 1 1 1 1 1 1
#> Levels: 1

# Specifying number of cells, features and samples
l2 = create_scDataset_raw(cells = 500, features = 500, nsamp=2)

# Specifying species
mouse_l = create_scDataset_raw(ref="mm10")

# Specifying batches
batch_l = create_scDataset_raw(nsamp=4, batch_id = factor(c(1,1,2,2)))

# Peaks of different size as features
peak_l = create_scDataset_raw(featureType="peak")
head(peak_l$mat)

#> 6 x 300 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 300 column names ‘sample_1_c1’, ‘sample_1_c2’, ‘sample_1_c3’ ... ]]

#>                                                                               
#> chr1_5998583_6998346   . .   2 1 .   . .   .   1 .   .   2 1   .   1   . .   1
#> chr1_6250001_6260000   . . 419 1 1   1 .   . 178 1   .   1 2   .   1   . .   1
#> chr1_7720001_7721000   . 1   2 . .   1 .  74   . . 165   1 3 344   .   2 1   .
#> chr1_9300001_9302500   . 1 229 1 .   . .   1   4 6   . 238 .   .   .   . . 268
#> chr1_15646045_15654043 . .   . . . 270 1 151   . .   1   1 1   . 454 419 2   3
#> chr1_18995510_19995273 . .   . . .   . .   1   1 1   .   . .   1   .   . .   .
#>                                                                               
#> chr1_5998583_6998346    .   1 1 1   1 295 1 1   1 . 165 .   . .  . 153   . . 2
#> chr1_6250001_6260000    .   . . .   3   . 1 2   . . 191 .   2 .  .   . 144 . .
#> chr1_7720001_7721000    2   . . .   .   1 1 .  31 .   . 1   . .  .   2   . . 1
#> chr1_9300001_9302500    2   . 1 2   .   . 2 1   2 .   . .   1 2  .  31   2 . 3
#> chr1_15646045_15654043  1 264 . .   1 353 2 .   . .   . . 273 .  .   .   . 2 .
#> chr1_18995510_19995273 65   . . 1 237 120 . . 140 1   2 .   . 1 34   .   . . 2
#>                                                                                
#> chr1_5998583_6998346     1   . . 206 .   6 .   . .  2 . . 230   .   6   1 . 347
#> chr1_6250001_6260000     .   . .   . .   2 .   . .  2 . 1   . 116 181 236 .   .
#> chr1_7720001_7721000     1   1 .   . 1   1 .   . . 15 . 1   .   .   6   2 1   .
#> chr1_9300001_9302500     1   . 1   1 .   . .   1 .  . . 2 261   .   .   1 .   .
#> chr1_15646045_15654043   .   2 . 236 1   2 2 262 .  1 . . 352   2   2   . 1   .
#> chr1_18995510_19995273 182 194 .   . . 262 2   . .  . . 1   .   . 106   1 .   .
#>                                                                                
#> chr1_5998583_6998346     1   .   3 340 239   . 1 1   . 396 . 2 . 1 . . 19 .   .
#> chr1_6250001_6260000   271 216   1   .   . 173 . .   .   . . . 2 . . 1  . . 146
#> chr1_7720001_7721000     1   1 214   2   .   2 2 .   1 269 . . . 1 . .  1 .   .
#> chr1_9300001_9302500     2   .   1   1   .   . . .   .   . . 3 2 . 1 .  1 .   1
#> chr1_15646045_15654043   2   .   .   .   .   . . . 268   . 1 . 2 . 1 .  6 7   .
#> chr1_18995510_19995273   .   .   2   .   .   . . .   .   1 . 1 1 . . 1  . .   .
#>                                                                               
#> chr1_5998583_6998346   120 2 .   . . . .   . 248 .   .   1 153 3   . 257   . .
#> chr1_6250001_6260000     . 1 .   2 1 . 1   .   . 1   .   2   . .   .   2   1 .
#> chr1_7720001_7721000     . . 6   . . . .   .   . 1   . 168 475 .   1   2   . .
#> chr1_9300001_9302500     . . .   3 2 . .   .   . .   .   .   . . 135   .   . 1
#> chr1_15646045_15654043   1 . .   . 1 . .   .   . 1 542   .   . 1   .   . 209 .
#> chr1_18995510_19995273 536 . . 319 1 4 . 132   1 1   .   1   . 1   6   .   . .
#>                                                                               
#> chr1_5998583_6998346     1   . . .   .   1 128 .   . .   . . 1 .   .   . . . 1
#> chr1_6250001_6260000   120 370 . .   .   .   . .   . .   . 1 1 . 130   1 . . .
#> chr1_7720001_7721000     2   . 2 4 309   1   . 1   . .   . 2 . .   .   . . 1 .
#> chr1_9300001_9302500     .   . . . 253 170 127 1   . .   . . 1 .   .   . . . .
#> chr1_15646045_15654043   .   . . .   .   1  31 1   . . 241 . . .   . 131 1 . 2
#> chr1_18995510_19995273   .   . 1 .   .   .   . . 156 .   . . . .   3   . . 2 1
#>                                                                               
#> chr1_5998583_6998346     . 253 . .   .   3 142 124 1 3 1  . . . .  1 . . 285 2
#> chr1_6250001_6260000     .   1 3 .   .  76   1   . . 1 .  . . 1 .  7 1 .   . 1
#> chr1_7720001_7721000     .   . . 1 290   .   2   . . . .  . . 1 .  1 . .   . .
#> chr1_9300001_9302500     1   4 . .   .   1   .   . . . .  2 . . .  . . .   . 1
#> chr1_15646045_15654043   .   . . .   1 273 139   1 . 2 . 84 . 1 . 67 2 1   . .
#> chr1_18995510_19995273 329   1 . .   .   . 208   . . . .  . . 3 .  . . .   3 .
#>                                                                                
#> chr1_5998583_6998346   .  3   1 .   .  1 .  2   2   . . 151 . . .   . 1   .   2
#> chr1_6250001_6260000   .  .   . .   1  . .  1   1 113 .   4 . . .   . . 361 269
#> chr1_7720001_7721000   .  .   . . 226  . . 72   .   . 2 367 . . . 394 .   .   .
#> chr1_9300001_9302500   .  .   . 1   .  . 2  .   .   . .   . . 2 .   1 .   .   2
#> chr1_15646045_15654043 . 76 146 .   . 84 .  . 106   . .   . 1 . 6   . .   1   .
#> chr1_18995510_19995273 2  1   1 .   .  . .  .   .   3 1   . 1 . .   . .   .   .
#>                                                                                
#> chr1_5998583_6998346   .   . . . 164   . .   .   1 . . . 2 .   . 1 .  . 2 1   .
#> chr1_6250001_6260000   . 304 . .   .   . .   1   . . . . . .   . . . 92 . 1   1
#> chr1_7720001_7721000   1   . . .   .   . .   1 164 . . . 2 .   1 . 2  . . 1   .
#> chr1_9300001_9302500   .   . 3 .   .   2 . 252   2 . 1 . . .   1 . .  . . .   .
#> chr1_15646045_15654043 .   . . .   1   1 .   .   1 3 . . 2 . 156 . .  . . 1   3
#> chr1_18995510_19995273 .   . . .   . 116 .   3   1 2 . . 3 1 177 . .  3 . . 105
#>                                                                               
#> chr1_5998583_6998346   1   1   . 430  .   3 3 .   .   . . . . 42 . . .   1 313
#> chr1_6250001_6260000   2   .   1   2  1   1 . . 252   . . . .  1 . 1 1   1   1
#> chr1_7720001_7721000   1   .   .   .  1 229 . .   .   2 . . .  . . . .   .   .
#> chr1_9300001_9302500   . 107 175   1 80   1 1 .   1   . . . .  . 2 . . 384   1
#> chr1_15646045_15654043 .   .   1   .  .   . . .   .   . . 2 .  1 . . 1   1   .
#> chr1_18995510_19995273 1   .   .   2  .   . 1 1   . 281 . . 1  . . . 2   .   .
#>                                                                                
#> chr1_5998583_6998346   2 . .   1 .   . 197 . 196 1 . .   . 237 .   . . . . .  .
#> chr1_6250001_6260000   . . .   . .   .   3 .   1 . 2 .   1   . .   . . . 1 1  .
#> chr1_7720001_7721000   . 1 .   . .   .   . .   . . 1 1   .   . .   . . . . .  2
#> chr1_9300001_9302500   2 . .   . .   .   2 .   . 1 1 1   .   3 . 404 . 1 . 1  .
#> chr1_15646045_15654043 . . 2 161 .   .   . .   3 1 1 1   .   . .   1 . . 2 . 37
#> chr1_18995510_19995273 . . .   . 1 246   . . 180 . . 1 236   2 1   . . 3 . .  .
#>                                                                             
#> chr1_5998583_6998346   . .   3 2 . . 1 . 143 160   . . . 2  1   . .   1 2  1
#> chr1_6250001_6260000   . .   . 1 6 . . .   1   .   . 2 . 2  .   . .   2 4  .
#> chr1_7720001_7721000   2 .   1 . 3 1 . 3   .   .   . . . .  .   1 .   1 1  .
#> chr1_9300001_9302500   1 . 347 1 . . . .   1   . 156 1 . .  3   . .   . .  3
#> chr1_15646045_15654043 . 1   . 2 . . . .   .   .   . . 1 1 60   . 1 252 . 32
#> chr1_18995510_19995273 1 .   1 . . 1 . 1   2  62 234 . . .  . 175 .   . .  2
#>                                                                               
#> chr1_5998583_6998346     .   . .   . . .   1   1 .   . .   1 296   2  . 1 1  .
#> chr1_6250001_6260000     .   . .   . 2 .   2   . .   . .   .   1   . 47 1 . 79
#> chr1_7720001_7721000   306 191 .   . . .   . 250 . 303 . 259   1 180  1 . .  3
#> chr1_9300001_9302500     . 255 6   . 2 1   .   . 2   . 1   .   .   .  6 . .  1
#> chr1_15646045_15654043   .   3 . 234 . 2 110   . 2  20 .   .   .   2  . 1 1  .
#> chr1_18995510_19995273   .   . .   1 . .   .   . .   . .   . 328   .  7 . .  .
#>                                                                              
#> chr1_5998583_6998346   . . 1 . 346 1   . . 2 210   .   1 . . .   .  . 1 . 1 1
#> chr1_6250001_6260000   . . . .   1 .   . . .   . 332  45 . 1 1   .  2 . . . .
#> chr1_7720001_7721000   . 2 1 1   2 .   1 . .   .   .   2 . . 1 105  . . 2 . .
#> chr1_9300001_9302500   . . 2 .   1 .   . 1 .   .   .   . . . .   .  2 . 1 . .
#> chr1_15646045_15654043 . 2 . 1   . . 338 . .   .   . 201 . . .   .  1 1 . . .
#> chr1_18995510_19995273 . . . .   . .   . . 3   .   1   . 2 . 1   . 71 . . . 1
#>                                                                                
#> chr1_5998583_6998346   167 . 105   .   . 63   3   .   1 257 . . 1 353 . 1 1 2 1
#> chr1_6250001_6260000     1 1   2   .   .  1   .   .   .   1 2 . .   1 . . . . .
#> chr1_7720001_7721000     . .  87 116   .  1   .   .   .   3 2 1 .   1 . . . . .
#> chr1_9300001_9302500     . .   1   .   4  .   1 329 272   . 1 1 1   1 . 2 . . 1
#> chr1_15646045_15654043 297 . 314   . 197  1 237   .   1   . . 2 .   . . . . . .
#> chr1_18995510_19995273   1 .   .   .   .  .   2   .   .   . . 3 .   . . . . . 2
#>                                                     
#> chr1_5998583_6998346   . . 2   1 . . 177   1 .   1 1
#> chr1_6250001_6260000   1 1 2 275 2 . 100   1 .   2 .
#> chr1_7720001_7721000   . . .   1 1 .   .   . 2   . 1
#> chr1_9300001_9302500   . . 1   . . .   2   . 1   . .
#> chr1_15646045_15654043 . 1 .   . . .   .   . 1   2 1
#> chr1_18995510_19995273 . 1 .   . 1 .   . 139 . 163 6

# Genes as features
gene_l = create_scDataset_raw(featureType="gene")
head(gene_l$mat)

#> 6 x 300 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 300 column names ‘sample_1_c1’, ‘sample_1_c2’, ‘sample_1_c3’ ... ]]

#>                                                                               
#> RP11-526P5.1    . . . 2 1 . 1 . .   . 41 111   1   2 2 . . 1 1 . 1 . . . .   1
#> CTD-2659N19.2   . 3 1 . . 2 . . . 301  .   . 149 128 . . . 2 . . . . . . 1   1
#> RP11-344P13.1 117 . . . . . 2 . .   1  .   . 284   . . 1 . . . . . 1 1 1 3   .
#> BARX1-AS1       . 1 . 3 . . . . 1   .  1   .   . 132 1 . . . 1 . . . 1 . .  56
#> PLAC1           1 . 1 1 2 1 . 1 .   .  2 266   .   . . . . . 2 . 1 6 . . 2 265
#> RP1-69D17.3     2 . 2 3 2 . . 1 .   .  .   .  89  59 . . . . . 1 1 . . . 1   .
#>                                                                               
#> RP11-526P5.1  . 1 . 250   . . .   6 . .   6 1 110   1 1   . . . 1  .   1   1 1
#> CTD-2659N19.2 . . .   .   2 . .   . . 1   . .   .   1 .   . . 2 1 88 119   . .
#> RP11-344P13.1 . 3 2   1   2 . .   1 . . 129 .   .   . . 155 2 . .  1   1   . .
#> BARX1-AS1     . . .   1   . 3 . 216 . .   . .   .   . .   . 1 . .  . 217   . 1
#> PLAC1         2 . .   .   1 . .   3 2 . 185 .   .   . . 222 . 1 2  1   2 138 .
#> RP1-69D17.3   . 3 .   . 274 . .   . . .   . 1   . 208 .   . 1 . 1  .   .   . 6
#>                                                                                
#> RP11-526P5.1  358   . 1 294 131  . . 2   .  . .   4 1 . 130 178 . 1 .   .   1 .
#> CTD-2659N19.2   .   . .   .   .  2 . .   .  . .   2 . .   .   1 . . .   .  78 .
#> RP11-344P13.1   3 145 .   1   .  . . .   . 47 .   . . 2   .   . 2 . .   1   . .
#> BARX1-AS1       1   . .   .   . 78 1 .   .  . .   1 . .   .   2 . . 1 340   . 1
#> PLAC1           .   . 2   .   1  1 . 2 346  . 1   . . 1   .   1 . . . 240   . .
#> RP1-69D17.3     .   . .   .   .  . 1 3   .  . . 216 1 1   1   . 1 . .   . 200 1
#>                                                                                
#> RP11-526P5.1    . .   . . 117 .   1  . . 129  2 1 2  .   . 1 . . .   .   .  1 .
#> CTD-2659N19.2   1 1   1 .   . .   .  . .   1  . . . 82   . 1 . 1 . 256   1  . .
#> RP11-344P13.1   3 .   . .   . 1   2  . .   . 94 2 1  . 116 . . 1 .   .   .  . 1
#> BARX1-AS1       1 1 367 1   1 .   .  . .   .  . 1 .  .   . . 1 3 .   .   1  1 .
#> PLAC1         109 1   . .   . . 185 78 1   1  1 1 1  .   2 . . . 2   . 121 79 2
#> RP1-69D17.3     2 3   . 2   . .   1  2 .   .  1 . .  1   2 . 1 . 1   . 154  . .
#>                                                                               
#> RP11-526P5.1    .   1 .   .   1 . .  . 106 .   .   . 1   . . .   .   .  . 2  .
#> CTD-2659N19.2   .   2 .   2 110 . 1  3   1 .   .   2 . 190 . .   .   .  . 1  3
#> RP11-344P13.1   .   . . 139   1 . 2  2   . 1   .   . .   . 1 .   . 205  . 2  .
#> BARX1-AS1     120   . .   .   1 . 1 81   . .   1   . .   1 . 2 161   2  . 1 86
#> PLAC1           . 267 1   .   1 2 .  .   . . 145 219 1   . . 1   .   . 43 .  .
#> RP1-69D17.3     .   . 2   1   . . .  .  22 .   .   . .   . 1 .   .   7  . .  1
#>                                                                               
#> RP11-526P5.1    1 .   . . 1 2 103 . . 221 .   2  .   .   .   . . 289   . . . .
#> CTD-2659N19.2 286 .   . . 2 .   . . .   . .   . 15 244   2   . .   1   . . . .
#> RP11-344P13.1   2 3   1 . . . 103 . 2   . .   .  .   3   . 326 2   . 331 1 . .
#> BARX1-AS1       . .   1 . . .   2 . .   1 .   .  1   2   . 119 .   .   . . . .
#> PLAC1           . 1   . . . .   . . .   . . 281  .   1 240   . .   .   . . 3 .
#> RP1-69D17.3     . 1 426 . . .   . . .   2 .   .  .   .   .   1 .   .   1 . . 1
#>                                                                               
#> RP11-526P5.1  . 1 .   . . 1 .   1   2 259 . . . 71 .   . 174   2 . 1   .   . .
#> CTD-2659N19.2 . . .   . . . 2   2   1   . 1 2 .  2 .   . 427   . 2 1 253   . .
#> RP11-344P13.1 2 1 .   1 . 1 . 195 231   . . . 1  2 .   .   1   1 . .   .   2 .
#> BARX1-AS1     . . .   . . . 6   .   .   . 3 . . 46 . 344   .   . . .   .   . 2
#> PLAC1         . . . 254 . . .   3   .   1 1 . .  . .  46   . 154 . .   .   . .
#> RP1-69D17.3   . . .   1 1 2 .   1 271   1 . . .  . .   .   1   . . 1   . 278 1
#>                                                                               
#> RP11-526P5.1    . 3 .  1 274 1   . . 1 2 1   3   . 1 1 . . 1   .   . . 2 1 . 1
#> CTD-2659N19.2   . . .  .   . .   . . 1 1 .   .   . 1 2 . 1 . 258   1 2 . . . .
#> RP11-344P13.1   . 2 .  .   . .   . . . . . 200   . . . . 1 3   3   1 . . 6 . .
#> BARX1-AS1       . . . 54   . . 223 . 1 4 .   . 275 . . . . 3   . 213 . 5 . . .
#> PLAC1           . 1 . 62   . .   . . . 1 . 171   . 2 1 . 4 .   .   1 1 . 1 1 .
#> RP1-69D17.3   167 1 .  .   1 .   . . . . .   . 148 . 2 . 1 .   .   . 1 1 . 1 2
#>                                                                               
#> RP11-526P5.1   87 . . 1   .   . .   . . .   . . .  .   2 . 1 390 . .   . .   1
#> CTD-2659N19.2 240 . 1 3   1 273 . 161 1 .   2 . . 64   . . .   . 2 .   . .   3
#> RP11-344P13.1 423 . . 3 290   . .   2 . .   . 1 .  1   1 . .   1 1 1   . . 187
#> BARX1-AS1       1 . . 2   1   1 2   2 . .   1 . .  1   . . .   1 . . 228 .   .
#> PLAC1           . . 1 .   .   . .   . 1 1   . . .  . 128 1 .   . . 2 318 .   .
#> RP1-69D17.3     . . . 1 336   . 2   . 1 . 466 . .  .   . 1 .   1 1 2   . .   1
#>                                                                               
#> RP11-526P5.1    1 1   2 . 176 2 3   1  87   . .   3 391 1   . . 3   1 . 1  . 3
#> CTD-2659N19.2 128 .  46 .   2 1 .   1 159   . 1   1   . .   2 2 .   . . 1  . .
#> RP11-344P13.1   1 .   . .   2 . 1   .   .   1 .   . 263 1 106 . 1   . . . 42 .
#> BARX1-AS1       1 1 253 .   2 3 1 135   1   . . 185   1 .   1 . .   2 . .  . .
#> PLAC1           2 .   . 3   2 . . 423 118   . .   .   . .   2 . . 128 2 .  . 1
#> RP1-69D17.3     . .   . .   1 . .   .   . 186 .   . 125 .   4 . .   . 1 .  . .
#>                                                                                
#> RP11-526P5.1  1 .   .   . .   3  .   . 1   1 2  . 382 . 1 . 134 1   1 .   .   1
#> CTD-2659N19.2 1 2 193   . .   1  .   . .   . 1 51   . . . 2   2 .   . . 212   .
#> RP11-344P13.1 . 1   .   . .   .  .   2 .   . 2  .   . 2 . .   . 1   . .   . 296
#> BARX1-AS1     . .   .   3 1   2 18   . . 202 .  .   1 . . 1   . . 439 7   .   2
#> PLAC1         . .   6 342 .   1  2 161 1   2 .  1   . . . .   . .   . .   . 172
#> RP1-69D17.3   2 . 435   . 1 285  .   . 1   1 1  .   . 1 . .   1 .   . .   1   .
#>                                                                                
#> RP11-526P5.1    1 .  .   1 . . .  . 1 . . 312 .   1 . . 222 .   2 .   .   . . 1
#> CTD-2659N19.2   . .  .   1 1 . 2  . . . .   . .   . . .   . 1   1 .   1   2 . .
#> RP11-344P13.1   2 2 37   2 . 1 3 51 . . .   1 1   2 1 .   . . 201 . 202   . . 1
#> BARX1-AS1     181 1  .   . 2 . .  2 . . 2   . .   . . 2   . .   . 1   . 336 . .
#> PLAC1           . .  . 274 1 . .  1 7 1 .   . 1   1 2 1   2 .   . 3   .   2 . 2
#> RP1-69D17.3     1 .  1   1 . . .  1 . . .   6 . 212 . 1   . .   . .   .   . 2 1
#>                                                                             
#> RP11-526P5.1  . 2   . 86 142   . . 3 1 . 113 . . .   . 1 1 .   . . 1 242 . 1
#> CTD-2659N19.2 . .   .  .   . 200 . 1 . .   2 . . .   . . . .   . . .   . . 1
#> RP11-344P13.1 . .   .  1   . 171 1 . . .   1 . . .   . 1 . . 257 . 2   . . .
#> BARX1-AS1     2 2   .  1 371   . 1 . . .   . 1 . .   1 2 . .   . . .   . . 1
#> PLAC1         1 2 201  3   .   1 1 . . .   . 1 1 .   2 . . 1   1 . .   . . 1
#> RP1-69D17.3   . .   .  .   .   . . . . .   1 . . . 103 . 1 4   . . .   1 . .